document_converter/requirements.txt
# Document Classification and Conversion System - Requirements

# OCR Libraries
paddleocr>=2.7.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0

# PDF Processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Excel Generation
openpyxl>=3.1.0

# Machine Learning
scikit-learn>=1.3.0

# Image Processing (for OCR preprocessing)
opencv-python>=4.8.0

# GUI (tkinter is built-in, but PySimpleGUI is an alternative)
# PySimpleGUI>=4.60.0  # Optional - uncomment if preferred over tkinter

